Introduce _<part>_idim variable. Prepare for use with laserscad.
idim standands for internal dimension (the dimension lasercut library works on).
Later there will also be xdim which is the dimension of the actual parts on the
sheet. This accounts for the extra space used by the joint fingers, etc. The
laserscad script requires this to determine the bounding box for each
component. Also the naming scheme can (hopefully) reduce potential confusion
that may be caused by BOX_SIZE since BOX_SIZE mapping is IMO not as
straightforward as pure xy pairs especially when there's a lot of code
involved.